+++ src/sampletones/exporters/triangle.py
@@ -47,7 +47,7 @@ class TriangleExporter(Exporter[TriangleInstruction]):
 
         return {
             FeatureKey.INITIAL_PITCH: initial_pitch,
-            FeatureKey.VOLUME: np.array(volumes, dtype=np.int8),
-            FeatureKey.ARPEGGIO: np.array(np.diff(np.array(pitches)) % 16, dtype=np.int8),
-            FeatureKey.HI_PITCH: np.array(np.diff(np.array(pitches)) // 16, dtype=np.int8),
+            FeatureKey.VOLUME: np.array(volumes),
+            FeatureKey.ARPEGGIO: np.array(np.diff(np.array(pitches)) % 16),
+            FeatureKey.HI_PITCH: np.array(np.diff(np.array(pitches)) // 16),
         }
